backend/requirements.txt
fastapi==0.131.0
sqlalchemy==2.0.46
pydantic==2.12.5
psycopg2-binary==2.9.11
passlib==1.7.4
bcrypt==4.0.1
uvicorn==0.41.0
python-dotenv==1.2.1
email-validator==2.3.0
PyJWT==2.10.1
